+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/core.js
@@ -1,5 +1,5 @@
 // -*- coding: utf-8 -*-
-// (c) 2013,2014 Andreas Motl, Elmyra UG
+// (c) 2013-2016 Andreas Motl, Elmyra UG
 
 function getconfig(name, options) {
     options = options || {};
@@ -40,38 +40,4 @@ function propagate_opaque_errors() {
 
         });
     }
-}
-
-function boot_application() {
-
-    console.log('boot_application');
-
-    // initialize content which still resides on page level (i.e. no template yet)
-    $('#query').val(opsChooserApp.config.get('query'));
-    $('#ui-title').html(getconfig('setting.ui.page.title'));
-    $('#ui-subtitle').html(getconfig('setting.ui.page.subtitle'));
-    $('#ui-statusline').html(getconfig('setting.ui.page.statusline'));
-    $('#ui-productname').html(getconfig('setting.ui.productname'));
-    $('#ui-footer').html(getconfig('setting.ui.page.footer', {after: '<br/>'}));
-    $('#ui-footer-version').html(getconfig('setting.ui.version', {after: '<br/>'}));
-
-    // propagate "datasource" query parameter
-    var datasource = opsChooserApp.config.get('datasource');
-    if (datasource) {
-        opsChooserApp.set_datasource(datasource);
-    }
-
-    // activate regular list region right now to avoid double rendering on initial page load
-    opsChooserApp.listRegion.show(opsChooserApp.collectionView);
-
-    // hide pagination- and metadata-area to start from scratch
-    opsChooserApp.ui.reset_content();
-
-    // propagate opaque error messages to alert area
-    propagate_opaque_errors();
-
-    opsChooserApp.trigger('application:ready');
-
-    opsChooserApp.ui.do_element_visibility();
-
 }