+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/boot/embedded.js
@@ -97,7 +97,7 @@ $(document).ready(function() {
     });
 
     opsChooserApp.start();
-    boot_application();
+    opsChooserApp.trigger('application:boot');
 
     // Automatically run search after bootstrapping application.
     // However, from now on [2014-05-21] this gets triggered by "project:ready" events.